requirements.txt
google-generativeai
Pillow
python-dotenv
telegram
